refactor: 패키지 구조 변경 (#85)
Intro(도입화면) 패키지 생성
