+++ backend/src/main/java/com/example/backend/domain/information/controller/InformationController.java
@@ -4,7 +4,6 @@ import com.example.backend.domain.global.dto.ResponseDto;
 import com.example.backend.domain.global.model.enums.ErrorCode;
 import com.example.backend.domain.information.dto.CommonResponse;
 import com.example.backend.domain.information.service.InformationStrategyFactory;
-//import com.example.backend.domain.information.service.OptionInformationService;
 import com.example.backend.domain.information.service.strategy.InformationStrategy;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.annotation.Order;
@@ -18,9 +17,7 @@ import java.util.List;
 @RequestMapping("/api/v1/info")
 @RequiredArgsConstructor
 public class InformationController {
-    public static final String ADDITIONAL_OPTION = "additional_option";
     private final InformationStrategyFactory strategyFactory;
-//    private final OptionInformationService optionInformationService;
 
     @Order(2)
     @GetMapping("/{targetInfo}")
