+++ backend/src/main/java/com/example/backend/domain/information/model/car/entity/Trim.java
@@ -1,6 +1,7 @@
 package com.example.backend.domain.information.model.car.entity;
 
 import com.example.backend.domain.global.model.BaseInfo;
+import com.example.backend.domain.intro.entity.TrimExterior;
 import lombok.Builder;
 import lombok.Getter;
 import org.springframework.data.annotation.Id;
