+++ backend/src/main/java/com/example/backend/domain/intro/dto/IntroResponse.java
@@ -1,4 +1,4 @@
-package com.example.backend.domain.information.dto;
+package com.example.backend.domain.intro.dto;
 
 import lombok.Getter;
