+++ backend/src/main/java/com/example/backend/domain/intro/entity/TrimInterior.java
@@ -1,5 +1,7 @@
-package com.example.backend.domain.information.model.car.entity;
+package com.example.backend.domain.intro.entity;
 
+import com.example.backend.domain.information.model.car.entity.InteriorColor;
+import com.example.backend.domain.information.model.car.entity.Trim;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.Transient;
 import org.springframework.data.relational.core.mapping.Table;
