+++ backend/src/main/java/com/example/backend/domain/intro/mapper/TrimInteriorRowMapper.java
@@ -1,8 +1,8 @@
-package com.example.backend.domain.information.model.car.repository;
+package com.example.backend.domain.intro.mapper;
 
 import com.example.backend.domain.information.model.car.entity.InteriorColor;
 import com.example.backend.domain.information.model.car.entity.Trim;
-import com.example.backend.domain.information.model.car.entity.TrimInterior;
+import com.example.backend.domain.intro.entity.TrimInterior;
 import org.springframework.jdbc.core.RowMapper;
 
 import java.sql.ResultSet;
