+++ backend/src/main/java/com/example/backend/domain/intro/repository/TrimAdditionalOptionRepository.java
@@ -1,6 +1,7 @@
-package com.example.backend.domain.information.model.car.repository;
+package com.example.backend.domain.intro.repository;
 
-import com.example.backend.domain.information.model.car.entity.TrimAdditionalOption;
+import com.example.backend.domain.intro.entity.TrimAdditionalOption;
+import com.example.backend.domain.intro.mapper.TrimAdditionalOptionRowMapper;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
