+++ backend/src/main/java/com/example/backend/domain/intro/service/OptionInformationService.java
@@ -1,13 +1,12 @@
-package com.example.backend.domain.information.service;
+package com.example.backend.domain.intro.service;
 
 import com.example.backend.domain.global.exception.RestApiException;
 import com.example.backend.domain.global.model.enums.ErrorCode;
-import com.example.backend.domain.information.dto.IntroResponse;
-import com.example.backend.domain.information.dto.SimpleIntroResponse;
-import com.example.backend.domain.information.mapper.IntroSourceMapper;
-import com.example.backend.domain.information.model.car.entity.TrimAdditionalOption;
-import com.example.backend.domain.information.model.car.repository.TrimAdditionalOptionRepository;
-import com.example.backend.domain.information.model.option.repository.AdditionalOptionRepository;
+import com.example.backend.domain.intro.dto.IntroResponse;
+import com.example.backend.domain.intro.dto.SimpleIntroResponse;
+import com.example.backend.domain.intro.mapper.IntroSourceMapper;
+import com.example.backend.domain.intro.entity.TrimAdditionalOption;
+import com.example.backend.domain.intro.repository.TrimAdditionalOptionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
@@ -18,7 +17,6 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class OptionInformationService {
-    private final AdditionalOptionRepository optionRepository;
     private final IntroSourceMapper introSourceMapper;
     private final TrimAdditionalOptionRepository trimOptionRepository;
 
@@ -28,7 +26,6 @@ public class OptionInformationService {
         if (size == -1) throw new RestApiException(ErrorCode.PAGE_SIZE_NOT_FOUND);
 
         PageRequest pageRequest = PageRequest.of(page, size);
-//        if(category.isBlank()) return findWholeOptions(pageRequest);
 
         return findAllByCategory(trimId, category, pageRequest);
     }
